refactored components, moved all api calls into separate component
