+++ src/components/Gif.js
@@ -1,7 +1,7 @@
-import axios from 'axios';
 import { useEffect, useState } from 'react';
 import AOS from 'aos';
 import 'aos/dist/aos.css'; 
+import { getGif } from "../components/ApiCalls.js"
 AOS.init();
 
 function Gif(props) {
@@ -10,16 +10,7 @@ function Gif(props) {
 	console.log('I just rendered with these keywords', props.keyword);
 
 	useEffect(() => {
-		axios({
-			url: `https://api.giphy.com/v1/gifs/search`,
-			params: {
-				api_key: `NFjbXVR8Fnr6sKnvC2hgL2etOmY2z7hO`,
-				q: props.keyword,
-				limit: 1,
-			},
-		}).then((res) => {
-			setCurrentGifs(res.data.data);
-		});
+		getGif(props, setCurrentGifs)
 	}, [props.keyword]);
 
 	console.log('after setting current gifs', currentGifs);
