+++ src/components/SearchBar.js
@@ -1,59 +1,19 @@
 import play from '../assets/play.png';
 import { useEffect, useState } from 'react';
-import axios from 'axios';
-import Swal from 'sweetalert2';
+import { getMovieId } from './ApiCalls';
 
 const SearchBar = (props) => {
 	// state that holds the user's form input
 	const [input, setInput] = useState('');
-	const [savedInput, setSavedInput] = useState('welcome');
+	const [savedInput, setSavedInput] = useState('');
+
 
 	useEffect(() => {
-		try {
-			axios({
-				url: 'https://api.themoviedb.org/3/search/movie',
-				method: 'GET',
-				dataResponse: 'json',
-				params: {
-					api_key: props.apiKey,
-					query: savedInput,
-					language: 'en-US',
-				},
-			}).then((res) => {
-				console.log('first api call', res.data.results[0]);
-				if (res.data.results[0] === undefined) {
-					
-					Swal.fire({
-						icon: 'error',
-						title: 'This movie does not exist, please check the spelling and try again!',
-						background: '#B3D5DF',
-						color: '#616d6b',
-						borderRadius: '1',
-						showConfirmButton: false,
-						timer: 3000
-					})
-					
-				} else {
-					props.setId(res.data.results[0].id);
-					console.log("response", res)
-				}
-			}).catch(error => {
-				console.log("error handling", error)
-				if (error.response.status >= 500) {
-					alert("server error")
-					console.log("server error")
-				} else if (error.response.status > 300) {
-					alert("error, please try again")
-					console.log("error, please try again")
-				}
-	
-			});
+			if (props.id || savedInput) { 
+				getMovieId(props, savedInput) 
+			}
+	}, [savedInput]) 
 
-		} catch (error) {
-
-		}
-		// eslint-disable-next-line
-	}, [savedInput]);
 
 	const userChoice = (e) => {
 		setInput(e.target.value);
@@ -78,7 +38,7 @@ const SearchBar = (props) => {
 					type="text"
 					id="userMovieChoice"
 					value={input}
-          			required
+          required
 				/>
 
 				<button type="submit">
